requirements.txt
matplotlib==3.10.3
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
plotly==6.1.1
seaborn==0.13.2
streamlit==1.45.1
